+++ modules/vaos/spec/controllers/v2/referrals_controller_spec.rb
@@ -321,7 +321,8 @@ RSpec.describe VAOS::V2::ReferralsController, type: :request do
         allow_any_instance_of(Ccra::ReferralService).to receive(:get_referral)
           .with(referral_consult_id, icn)
           .and_return(referral_detail)
-        allow_any_instance_of(VAOS::V2::AppointmentsService).to receive(:get_active_appointments_for_referral)
+        allow_any_instance_of(VAOS::V2::CommunityCare::AppointmentCoordinator)
+          .to receive(:appointments_for_referral)
           .with(referral_number)
           .and_return({ EPS: { data: [] }, VAOS: { data: [] } })
       end
@@ -350,7 +351,8 @@ RSpec.describe VAOS::V2::ReferralsController, type: :request do
 
       context 'when EPS and VAOS have active appointments' do
         before do
-          allow_any_instance_of(VAOS::V2::AppointmentsService).to receive(:get_active_appointments_for_referral)
+          allow_any_instance_of(VAOS::V2::CommunityCare::AppointmentCoordinator)
+            .to receive(:appointments_for_referral)
             .with(referral_number)
             .and_return({
                           EPS: {
@@ -457,7 +459,8 @@ RSpec.describe VAOS::V2::ReferralsController, type: :request do
           allow_any_instance_of(Ccra::ReferralService).to receive(:get_referral)
             .with(referral_consult_id, icn)
             .and_return(referral_detail_missing_data)
-          allow_any_instance_of(VAOS::V2::AppointmentsService).to receive(:get_active_appointments_for_referral)
+          allow_any_instance_of(VAOS::V2::CommunityCare::AppointmentCoordinator)
+            .to receive(:appointments_for_referral)
             .with(referral_number)
             .and_return({ EPS: { data: [] }, VAOS: { data: [] } })
         end
@@ -504,7 +507,8 @@ RSpec.describe VAOS::V2::ReferralsController, type: :request do
             allow_any_instance_of(Ccra::ReferralService).to receive(:get_referral)
               .with(referral_consult_id, icn)
               .and_return(referral_detail_partial_missing)
-            allow_any_instance_of(VAOS::V2::AppointmentsService).to receive(:get_active_appointments_for_referral)
+            allow_any_instance_of(VAOS::V2::CommunityCare::AppointmentCoordinator)
+              .to receive(:appointments_for_referral)
               .with(referral_number)
               .and_return({ EPS: { data: [] }, VAOS: { data: [] } })
           end
@@ -531,7 +535,8 @@ RSpec.describe VAOS::V2::ReferralsController, type: :request do
             allow_any_instance_of(Ccra::ReferralService).to receive(:get_referral)
               .with(referral_consult_id, icn)
               .and_return(referral_detail)
-            allow_any_instance_of(VAOS::V2::AppointmentsService).to receive(:get_active_appointments_for_referral)
+            allow_any_instance_of(VAOS::V2::CommunityCare::AppointmentCoordinator)
+              .to receive(:appointments_for_referral)
               .with(referral_number)
               .and_return({ EPS: { data: [] }, VAOS: { data: [] } })
           end
