+++ modules/vaos/spec/requests/vaos/v2/appointments_spec.rb
@@ -1245,7 +1245,7 @@ RSpec.describe 'VAOS::V2::Appointments', :skip_mvi, type: :request do
               allow(StatsD).to receive(:histogram).with(any_args)
 
               expect(StatsD).to receive(:increment).with(
-                described_class::APPT_CREATION_SUCCESS_METRIC,
+                VAOS::V2::CommunityCare::SubmitAppointment::APPT_CREATION_SUCCESS_METRIC,
                 tags: ['service:community_care_appointments', 'type_of_care:no_value']
               )
 
@@ -1263,7 +1263,7 @@ RSpec.describe 'VAOS::V2::Appointments', :skip_mvi, type: :request do
                            match_requests_on: %i[method path]) do
             VCR.use_cassette('vaos/v2/eps/post_submit_appointment_conflict',
                              match_requests_on: %i[method path body]) do
-              expect_metric_increment(described_class::APPT_CREATION_FAILURE_METRIC) do
+              expect_metric_increment(VAOS::V2::CommunityCare::SubmitAppointment::APPT_CREATION_FAILURE_METRIC) do
                 post '/vaos/v2/appointments/submit', params:, headers: inflection_header
               end
 
@@ -1318,10 +1318,10 @@ RSpec.describe 'VAOS::V2::Appointments', :skip_mvi, type: :request do
                 expect(response).to have_http_status(:created)
 
                 expect(StatsD).to have_received(:increment).with(
-                  described_class::APPT_CREATION_SUCCESS_METRIC,
+                  VAOS::V2::CommunityCare::SubmitAppointment::APPT_CREATION_SUCCESS_METRIC,
                   tags: ['service:community_care_appointments', 'type_of_care:CARDIOLOGY']
                 )
-                expect(StatsD).to have_received(:histogram).with(described_class::APPT_CREATION_DURATION_METRIC,
+                expect(StatsD).to have_received(:histogram).with(VAOS::V2::CommunityCare::SubmitAppointment::APPT_CREATION_DURATION_METRIC,
                                                                  kind_of(Numeric),
                                                                  tags: ['service:community_care_appointments'])
               end
@@ -1350,7 +1350,7 @@ RSpec.describe 'VAOS::V2::Appointments', :skip_mvi, type: :request do
 
               # Verify metric was logged with 'no_value'
               expect(StatsD).to have_received(:increment).with(
-                described_class::APPT_CREATION_SUCCESS_METRIC,
+                VAOS::V2::CommunityCare::SubmitAppointment::APPT_CREATION_SUCCESS_METRIC,
                 tags: ['service:community_care_appointments', 'type_of_care:no_value']
               )
             end
@@ -1362,7 +1362,7 @@ RSpec.describe 'VAOS::V2::Appointments', :skip_mvi, type: :request do
                            match_requests_on: %i[method path]) do
             VCR.use_cassette('vaos/v2/eps/post_submit_appointment_400',
                              match_requests_on: %i[method path]) do
-              expect_metric_increment(described_class::APPT_CREATION_FAILURE_METRIC) do
+              expect_metric_increment(VAOS::V2::CommunityCare::SubmitAppointment::APPT_CREATION_FAILURE_METRIC) do
                 post '/vaos/v2/appointments/submit', params: { ** params, phone_number: nil },
                                                      headers: inflection_header
               end
@@ -1388,7 +1388,7 @@ RSpec.describe 'VAOS::V2::Appointments', :skip_mvi, type: :request do
               allow(StatsD).to receive(:increment).with(any_args)
 
               expect(StatsD).to receive(:increment).with(
-                described_class::APPT_CREATION_FAILURE_METRIC,
+                VAOS::V2::CommunityCare::SubmitAppointment::APPT_CREATION_FAILURE_METRIC,
                 tags: ['service:community_care_appointments', 'type_of_care:no_value']
               )
